(build) Exclude FluentAssertions from CodeCoverage
By default, when uploading to coveralls.io, it will attempt to upload
all .cs files for "things" that it has inspected.  FluentAssertions is
referenced in the codebase, and therefore OpenCover will trying to
include it.  Adding this to the array of exclusions, means that it
won't be inspected going forward.